Extract data methods to Infrastructure project
