+++ csharp/Bargreen.sln
@@ -3,11 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29411.108
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bargreen.Services", "src\Bargreen.Services\Bargreen.Services.csproj", "{AE12F9D4-1066-4C8A-9A0D-062C011191B6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bargreen.Services", "src\Bargreen.Services\Bargreen.Services.csproj", "{AE12F9D4-1066-4C8A-9A0D-062C011191B6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bargreen.Tests", "src\Bargreen.Tests\Bargreen.Tests.csproj", "{1E64FEE6-A227-444E-B7C1-7CC081163DEA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bargreen.Tests", "src\Bargreen.Tests\Bargreen.Tests.csproj", "{1E64FEE6-A227-444E-B7C1-7CC081163DEA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bargreen.API", "src\Bargreen.API\Bargreen.API.csproj", "{2C9C0612-365E-4793-80A8-F7E361C7C10C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bargreen.API", "src\Bargreen.API\Bargreen.API.csproj", "{2C9C0612-365E-4793-80A8-F7E361C7C10C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bargreen.Infrastructure", "src\Bargreen.Infrastructure\Bargreen.Infrastructure\Bargreen.Infrastructure.csproj", "{97899037-F6B3-4AB9-872E-617BF98A8649}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{2C9C0612-365E-4793-80A8-F7E361C7C10C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2C9C0612-365E-4793-80A8-F7E361C7C10C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2C9C0612-365E-4793-80A8-F7E361C7C10C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{97899037-F6B3-4AB9-872E-617BF98A8649}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{97899037-F6B3-4AB9-872E-617BF98A8649}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{97899037-F6B3-4AB9-872E-617BF98A8649}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{97899037-F6B3-4AB9-872E-617BF98A8649}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
