+++ csharp/src/Bargreen.API/Bargreen.API.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
@@ -17,6 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Bargreen.Infrastructure\Bargreen.Infrastructure\Bargreen.Infrastructure.csproj" />
     <ProjectReference Include="..\Bargreen.Services\Bargreen.Services.csproj" />
   </ItemGroup>
