+++ csharp/src/Bargreen.Infrastructure/Bargreen.Infrastructure/Bargreen.Infrastructure.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Bargreen.Services\Bargreen.Services.csproj" />
+    <ProjectReference Include="..\..\Bargreen.Services\Bargreen.Services.csproj" />
+  </ItemGroup>
+
+</Project>
