+++ csharp/src/Bargreen.Services/Interfaces/IInventoryRepository.cs
@@ -1,0 +1,11 @@
+﻿using Bargreen.Services.Dtos;
+using System.Collections.Generic;
+
+namespace Bargreen.Services.Interfaces
+{
+    public interface IInventoryRepository
+    {
+        IEnumerable<AccountingBalance> GetAccountingBalances();
+        IEnumerable<InventoryBalance> GetInventoryBalances();
+    }
+}
