+++ csharp/src/Bargreen.Services/Interfaces/IInventoryService.cs
@@ -5,8 +5,6 @@ namespace Bargreen.Services.Interfaces
 {
     public interface IInventoryService
     {
-        IEnumerable<AccountingBalance> GetAccountingBalances();
-        IEnumerable<InventoryBalance> GetInventoryBalances();
         IEnumerable<InventoryReconciliationResult> ReconcileInventoryToAccounting(IEnumerable<InventoryBalance> inventoryBalances, IEnumerable<AccountingBalance> accountingBalances);
     }
 }
